functions/apify_to_gcs/requirements.txt
google-cloud-storage>=2.17.0
requests>=2.31.0
